Made comments and improved UI function
Added comments for Ricardo, Jon, and Joseph who are working with the
intent I pass from my set up screen to the board class. Also added
function such that when a user selects 0 victory points, the attack
toggle button disappears and does not allow the user to set attacks to
false. If I allowed this, the game would never end as users would never
be able to kill of one anothers' settlements.