Created Class for managing views on the game board
BoardObjectManager is a class that manages all of the views that are on
the game board screen (roads, soldiers, settlements, etc.) The presence
of this class greatly cleans up the code in board screen main fragment.
More optimizations for this class to come, including adjustments for
padding. Note: this does not take the place of the Board model class.
The sole purpose of this class is to manage View Elements, not the
backend.