Got the board scaling mostly working.
There is still work to do. I think i may have to adjust for more navbar
stuff? not sure, because board doesn’t center perfectly. Additionally,
I couldn’t get the measuring right on the buttons, so I still need to
figure that out, as well as laying them out. Background may not appear
because it is too large for the bitmap to decode. Need to figure out a
fix for this (other than splicing the image into parts)